requirements.txt
requests
RPi.GPIO ; platform_machine == "armv7l"
